browserify build + commonJS require + restructuring
