+++ package.json
@@ -25,13 +25,15 @@
     "grunt-contrib-clean": "~0.4.1",
     "grunt-contrib-coffee": "~0.7.0",
     "grunt-contrib-connect": "~0.3.0",
-    "grunt-contrib-concat": "~0.3.0",
     "grunt-contrib-jshint": "~0.6.0",
     "grunt-contrib-uglify": "~0.2.2",
     "grunt-contrib-watch": "~0.5.1",
+    "grunt-browserify": "~1.2.11",
     "grunt-istanbul-coverage": "0.0.1",
     "grunt-plato": "~0.2.0",
+    "grunt-simple-mocha": "~0.4.0",
     "lodash": "~2.2.1",
+    "async": "~0.2.9",
     "mustache": "~0.7.2",
     "chai": "~1.8.1",
     "chai-stats": "~0.3.0"
