+++ src/Tools/index.js
@@ -1,0 +1,3 @@
+var _ = require('underscore');
+
+_.extend(exports, require('./offset'));
