--- src/index.js
@@ -1,1 +1,0 @@
-var AudioKit = typeof module === "undefined" ? {} : module.exports;
