--- src/nodes/index.js
@@ -1,1 +1,0 @@
-AudioKit.Nodes = {};
