--- src/tools/index.js
@@ -1,1 +1,0 @@
-AudioKit.Tools = {};
